Fixed mismatched type warnings with 64bit integers and size_t

+++ gemcutter/Application/FileSystem.h
@@ -118,7 +118,7 @@ namespace gem
 		std::ifstream file;
 		std::string buffer;
 
-		unsigned currentPos = 0;
+		size_t currentPos = 0;
 		unsigned length = 0;
 	};
 }
